requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
elasticsearch[async]==8.13.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-generativeai==0.7.2
pdfplumber==0.11.0
python-docx==1.1.2
openpyxl==3.1.2
python-multipart==0.0.9
minio==7.2.7
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
greenlet==3.0.3
beautifulsoup4==4.12.3
numpy<2.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
sentence-transformers==3.0.1
PyJWT==2.8.0